Fix nomad deploy and extend to web and publishing
Update the nomad plan to:
* Fix the networking configuration
* Simplify the templating logic to remove need to download tar with
environment variable template in it
* Enable variable scale of deployment per environment using configs
from manifest
* Fix service registration and health checks

This change does not expose the `pprof` endpoint currently. If this is
desired then a further change will be required.

The config was also updated to enable the prometheus exporter and
specify the ports to ensure they will always match the nomad plan in
future.