Fix nomad service check
The nomad service check has a defect in v0.12.x whereby it is unable to
recognise the port labels from the group network stanza when the service
check is registered at the task level. This lead to the deployments of
the collector failing. This is [a known defect in
Nomad](https://github.com/hashicorp/nomad/issues/8780) and was fixed in
v1.0.

The change is a workaround by moving both the networking and the service
stanzas to the group level.