Use a static port for prometheus exporter
The prometheus exporter endpoint was using a dynamic port. Due to our
deployment that would have required too many ports to be opened up to
allow the prometheus scraping with the current deployment. Restricting
this to a static port will limit us to running one instance of the otel
collector per instance, but that should be ok.