Workaround Nomad for sytstem job static ports bug
The version of Nomad we are running has a bug whereby assigning static
ports to system jobs (which is a standard use case) was broken and
resulted in an error stating that the port label can not be found. This
is a known Nomad bug that can only be addressed by upgrading Nomad which
we are unable to do at present
(https://github.com/hashicorp/nomad/issues/8934). As such this
workaround deploys the job as a `service` rather than `system` job
meaning that we will need to manually configure the number of instances
of the collector load balancer deployed to match the number of nomad
client instances. This will only be a temporary workaround until the
Kubernetes migration is completed.