Workaround for DNS discovery issues
There was an issue with the DNS service discovery and the docker bridge
networking that we can not work through in the current timescales. This
is a temporary solution to allow us to run multiple copies of the
aggregators. While this approach makes it much harder to scale the
aggregators out if needed, it is not expected to live long as the EKS
work will replace this shortly anyway.